rest/requirements.txt
Django==2.2
djangorestframework==3.9.2
mysqlclient==1.4.2.post1
python-dotenv==0.10.1
pytz==2019.1
sqlparse==0.3.0
django-filter==2.1.0